requirements.txt
fastapi
uvicorn[standard]
httpx
pycryptodome
pydantic            # 已安装的就是 v2
pydantic-settings   # 新增
PyJWT